Added L-R balance to midi_creator_2
Added L-R balance to midi_creator_2. Now the parameter is a list of midi files instead of a single file. The channels for each midi will be for balances, and the different midi files will be used for the "volume trick". (Larger list = better volume trick, more chances of it to work well)